Carbon Emission Functionality Finalized / Modularized
