+++ src/css/styles.css
@@ -41,41 +41,51 @@ footer {
     width: 100%;
     height: 100vh; /* Fullscreen map */
 }
-.toggle-marker-button {
+
+/* General styling for the dropdown container */
+#dropdown-container {
     position: absolute;
-    top: 60px;
-    right: 10px;
-    z-index: 1000; /* Ensure the button appears above the map */
-    background-color: white;
-    border: 2px solid #4CAF50;
-    border-radius: 5px;
-    padding: 10px 20px;
+    top: 70px; /* Adjust based on the header height */
+    right: 20px; /* Distance from the left edge */
+    z-index: 1000; /* Ensure it stays on top of the map */
+    background-color: #fff; /* White background for the dropdown */
+    border: 1px solid #ccc; /* Light border */
+    border-radius: 5px; /* Rounded corners */
+    padding: 10px; /* Add padding for aesthetics */
+    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1); /* Subtle shadow */
+    font-family: Arial, sans-serif;
+}
+
+#world-emissions {
+    margin-bottom: 10px; /* Space between this text and the dropdown */
     font-size: 16px;
-    cursor: pointer;
+    font-weight: bold;
+    color: #333; /* Adjust for readability */
 }
 
-.toggle-shape-button {
-    position: absolute;
-    top: 110px;
-    right: 10px;
-    z-index: 1000; /* Ensure the button appears above the map */
-    background-color: white;
-    border: 2px solid #4CAF50;
-    border-radius: 5px;
-    padding: 10px 20px;
-    font-size: 16px;
-    cursor: pointer;
+
+/* Styling for the dropdown select element */
+#yearSelect {
+    width: 150px; /* Width of the dropdown */
+    padding: 5px; /* Padding for the dropdown */
+    font-size: 14px; /* Font size for readability */
+    border: 1px solid #ccc; /* Light border */
+    border-radius: 5px; /* Rounded corners */
+    background-color: #f9f9f9; /* Light gray background */
+    cursor: pointer; /* Change cursor to pointer */
 }
 
-.toggle-marker-button.active {
-    background-color: #4CAF50;
-    color: white;
+/* Hover and focus styles for the dropdown */
+#yearSelect:hover, #yearSelect:focus {
+    border-color: #007BFF; /* Blue border on hover or focus */
+    outline: none; /* Remove default outline */
 }
 
-.toggle-shape-button.active {
-    background-color: #4CAF50;
-    color: white;
+/* Optional styling for dropdown text alignment */
+#yearSelect option {
+    text-align: center; /* Center align the dropdown text */
 }
+
 
 
 .chart-container {
