--- src/js/tempCodeRunnerFile.js
@@ -1,2 +1,0 @@
-//var bounds = L.latLngBounds([[65.8747, -169.6289], [65.8747, -169.6289]]);
-//map.setMaxBounds(bounds);
